+++ django/applications/catmaid/static/js/init.js
@@ -115,6 +115,12 @@ var project;
     // Timeout reference for message updates if no websockets are available
     this._messageTimeout = undefined;
 
+    // General update WebSockets connection reference (if available)
+    this._updateSocket = undefined;
+
+    // The number of attempts of re-opening a closed socket
+    this._updateSocketRetries = 0;
+
     // Do periodic update checks
     window.setTimeout(CATMAID.Init.checkVersion, CATMAID.Init.CHECK_VERSION_TIMEOUT_INTERVAL);
 
@@ -781,7 +787,7 @@ var project;
     // The date of the last unread message
     var latest_message_date = null;
 
-    return function() {
+    return function(singleRequest) {
       CATMAID.fetch('messages/latestunreaddate')
         .then(function(data) {
           // If there is a newer latest message than we know of, get all
@@ -800,20 +806,108 @@ var project;
           CATMAID.statusBar.replaceLast('Unable to check for messages (network may be disconnected).');
         })
         .then((function() {
-          // Check again later
-          this._messageTimeout = window.setTimeout(CATMAID.client.check_messages,
-              MSG_TIMEOUT_INTERVAL);
+          if (!singleRequest) {
+            // Check again later
+            this._messageTimeout = window.setTimeout(CATMAID.client.check_messages,
+                MSG_TIMEOUT_INTERVAL);
+          }
         }).bind(this));
     };
   })();
 
   /**
-   * Try to setup websocket channels to the back-end server to avoid long
+   * Parse WebSockets messages and take appropriate action.
+   */
+  Client.prototype._handleWebsockMessage = function(message) {
+    var data = JSON.parse(message.data);
+    if (!data) {
+      throw new CATMAID.ValueError("Unexpected message format: " +
+          message.data);
+    }
+    var handler = CATMAID.Client.messageHandlers.get(data.event);
+    if (!handler) {
+      throw new CATMAID.ValueError("Unexpected message from server: " +
+          message.data);
+    }
+    handler(this, data.payload);
+  };
+
+  Client.messageHandlers = new Map([[
+        'new-message',
+        function(client, payload) {
+          CATMAID.msg("New message", payload.message_title);
+          client.check_messages(true);
+        }
+      ], [
+        'unknown',
+        function(message) {
+          var report = "An unknown message has been received" +
+              (message ? (": " + message) : "");
+          CATMAID.warn(report);
+          console.log(report);
+        }
+      ]]);
+
+  /**
+   * Try to setup WebSockets channels to the back-end server to avoid long
    * polling for message updates and more. If this is not possible, resort to
    * long polling.
    */
   Client.prototype.refreshBackChannel = function() {
-    this.check_messages();
+    if (Modernizr.websockets) {
+      // Close existing socket, if any
+      if (this._updateSocket) {
+        this._updateSocket.close();
+      }
+      // Create new socket
+      var url = 'ws://' + window.location.host + CATMAID.makeURL('/channels/updates/');
+      this._updateSocket = new WebSocket(url);
+
+      this._updateSocket.onopen = (function() {
+        console.log('WebSockets connection established.');
+        this._updateSocketRetries = 0;
+      }).bind(this);
+
+      // If errors happen (e.g. the back-end doesn't support WebSockets), fall
+      // back to long polling.
+      this._updateSocket.onerror = (function() {
+        console.log('There was an error establishing WebSockets connections. ' +
+            'Falling back to long polling.');
+        this.refreshLongPolling();
+        // Don't call clone handler after error
+        this._updateSocket.onclose = null;
+      }).bind(this);
+
+      // When a WebSockets connection is closed, try to re-open it three times
+      // before falling back to long polling.
+      this._updateSocket.onclose = (function() {
+        let maxRetries = 3;
+        if (this._updateSocketRetries >= maxRetries) {
+          console.log('WebSockets connection closed and maximum number of ' +
+            'retries reached, falling back to long polling.');
+          this.refreshLongPolling();
+        } else {
+          this._updateSocket = undefined;
+          ++this._updateSocketRetries;
+          console.log('WebSockets connection closed. Trying to re-open it in 3 second, retry ' +
+              this._updateSocketRetries + '/' + maxRetries + '.');
+          setTimeout(this.refreshBackChannel.bind(this), 3000);
+        }
+
+      }).bind(this);
+
+      this._updateSocket.onmessage = this._handleWebsockMessage.bind(this);
+
+      // Initial message update without long polling
+      this.check_messages(true);
+
+      // In case the socket was opened before the event listener was registered.
+      if (this._updateSocket.readyState == WebSocket.OPEN) {
+        this._updateSocket.onopen();
+      }
+    } else {
+      this.refreshLongPolling();
+    }
   };
 
   /**
@@ -821,6 +915,13 @@ var project;
    */
   Client.prototype.closeBackChannel = function() {
     if (this._messageTimeout) window.clearTimeout(this._messageTimeout);
+  };
+
+  /**
+   * Enable long polling update functions.
+   */
+  Client.prototype.refreshLongPolling = function() {
+    this.check_messages();
   };
 
   /**